fix: fix cd
